added backend data and formatting changes
added some of the csv files on the backend for reference. removed from app.py unused listening ports from when I thought. The argsParser was removed from frontend_fingerprinting as it is no longer called directly, and the localstorage was removed from identify.html, as it was not necessary, and the display_results function was changed as well as